Move platform-agnostic files 'FeedImageViewModel' and 'FeedImageViewPresenter' to 'EssentialFeed' target
